Add Dispatch

+++ giface/iconnection.go
@@ -9,5 +9,5 @@ type IConnection interface {
 	GetConnID() uint64
 	GetServer() IServer
 	Send([]byte)
-	SetState(uint16)
+	SetState(uint8)
 }
